LearningCircle/OD_Starter_VS_Python/OD_Starter_VS_Python/requirements.txt
ultralytics
torch
torchvision
pillow
tqdm
